# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor Appointment API",
        "description": "API Documentation for Doctor Appointment System",
        "contact": {
            "email": "ujjwal999adhikari@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Doctor Appointment API Server"
        }
    ],
    "paths": {
...
        "/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get all the schedule list of the current doctor",
                "operationId": "66e486a11c0890278f5328be62bce51e",
                "responses": {
                    "200": {
                        "description": "A specific schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Schedules"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Store a newly created schedule for a doctor",
                "description": "Store a doctor's schedule and return the stored data along with a success message",
                "operationId": "9a8fc6c84a98c5fad3940068fce37f8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "day",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "day": {
                                        "type": "string",
                                        "enum": [
                                            "Sunday",
                                            "Monday",
                                            "Tuesday",
                                            "Wednesday",
                                            "Thursday",
                                            "Friday",
                                            "Saturday"
                                        ],
                                        "example": "Monday"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "06:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "07:30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Schedule created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "schedule": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "doctor_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "day": {
                                                    "type": "string",
                                                    "example": "Monday"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "example": "09:00:00"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "example": "17:00:00"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Schedule created successfully"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "success"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "day": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The day field is required."
                                                    }
                                                },
                                                "start_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The start_time field is required."
                                                    }
                                                },
                                                "end_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The end_time field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Schedules": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "day": {
                        "type": "string",
                        "example": "Monday"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "17:00:00"
                    },
                    "doctor_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-27T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-27T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: `Bearer {token}`",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}